moving features to cr command

+++ src/calcit.rs
@@ -50,7 +50,7 @@ pub enum Calcit {
     sym: Arc<str>,
     info: Arc<CalcitSymbolInfo>,
     /// positions in the tree of Cirru
-    location: Option<Arc<Vec<u8>>>,
+    location: Option<Arc<Vec<u16>>>,
   },
   /// local variable
   Local(CalcitLocal),
@@ -879,7 +879,7 @@ impl CalcitErr {
 pub struct NodeLocation {
   pub ns: Arc<str>,
   pub def: Arc<str>,
-  pub coord: Arc<Vec<u8>>,
+  pub coord: Arc<Vec<u16>>,
 }
 
 impl From<NodeLocation> for Edn {
@@ -911,7 +911,7 @@ impl fmt::Display for NodeLocation {
 }
 
 impl NodeLocation {
-  pub fn new(ns: Arc<str>, def: Arc<str>, coord: Arc<Vec<u8>>) -> Self {
+  pub fn new(ns: Arc<str>, def: Arc<str>, coord: Arc<Vec<u16>>) -> Self {
     NodeLocation {
       ns,
       def,
